1. change expense type to int 2. add additional test items

+++ lib/chart.dart
@@ -3,7 +3,7 @@ import 'package:flutter_animation_progress_bar/flutter_animation_progress_bar.da
 
 class Chart extends StatelessWidget {
   final double progressValue;
-  final double expenses;
+  final int expenses;
   final String date;
 
   Chart({
